improved callendar views
