+++ app/src/view/calendarview.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 # @author Jakub Mazurkiewicz
 
+import datetime
 import tkinter as tk
 from tkcalendar import Calendar, DateEntry
 
@@ -9,13 +10,19 @@ from view import View
 class CalendarView(View):
     def __init__(self, parent):
         View.__init__(self, parent)
-        self.calendar = Calendar(self, selectmode='none')
-        self.button_frame = tk.Frame(self)
-        self.__build_grid()
+        
+        self.mindate = datetime.date.today() - datetime.timedelta(days = 30)
 
+        self.calendar = Calendar(parent, font="Comic_Sans 14", selectmode='day', locale='en_US',
+                   mindate=self.mindate, disabledforeground='red',
+                   cursor="hand2",width = 300, height = 400)
+
+        self.calendar.tag_config('reminder', background = 'red', foreground = 'blue')
+        self.button_frame = tk.Frame(parent)
+        self.__build_grid()
     
     def __build_grid(self):
         print('CalendarView::__build_grid')
-        self.calendar.grid(row=0, column=0)
-        self.button_frame.grid(row=1, column=0)
+        self.calendar.grid(column=0, row=0, columnspan = 7)
+        self.button_frame.grid(column=0, row=1)
