+++ app/src/view/employeesscheduleview.py
@@ -2,21 +2,20 @@
 # @author Jakub Mazurkiewicz
 
 import tkinter as tk
-from view import View
+from view.calendarview import CalendarView
 
-class EmployeesScheduleView(View):
+class EmployeesScheduleView(CalendarView):
     def __init__(self, parent):
-        View.__init__(self, parent)
+        CalendarView.__init__(self, parent)
 
-        self.show_order_location_button = tk.Button(self.button_frame, text='')
-        self.add_extra_costs_button = tk.Button(self.button_frame, text='')
+        self.show_order_location_button = tk.Button(self.button_frame, text='Show Order')
+        self.add_extra_costs_button = tk.Button(self.button_frame, text='Add extra costs')
 
         self.show_order_location_button.config(command=self.on_show_order_location_button_click)
         self.add_extra_costs_button.config(command=self.on_add_extra_costs_button_click)
 
         self.show_order_location_button.grid(row=0, column=0)
         self.add_extra_costs_button.grid(row=0, column=1)
-
 
     def on_show_order_location_button_click(self):
         print('show order')
