+++ app/src/view/mainwindow.py
@@ -6,6 +6,9 @@ from os import path
 
 from model import AuthorizationModel
 from view.authorizationview import AuthorizationView
+from view.orderscheduleview import OrderScheduleView
+from view.employeesscheduleview import EmployeesScheduleView
+from view.calendarview import CalendarView
 
 class MainWindow(tk.Tk):
     def __init__(self):
@@ -15,9 +18,12 @@ class MainWindow(tk.Tk):
 
         self.iconphoto(False, tk.PhotoImage(file='view/img/cheese.png'))
 
-        self.authorization_model = AuthorizationModel()
-        self.authorization_view = AuthorizationView(self)
-        self.authorization_view.set_model(self.authorization_model)
+        #self.authorization_model = AuthorizationModel()
+        #self.authorization_view = AuthorizationView(self)
+        #self.employees_schedule_view = EmployeesScheduleView(self)
+        self.order_schedule_view = OrderScheduleView(self)
+        
+        #self.authorization_view.set_model(self.authorization_model)
         
         self.grid_columnconfigure(0, weight=1)
         self.grid_rowconfigure(0, weight=1)
